git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK6@1692168 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/internal/jaxb/gco/GO_Integer.java
@@ -46,7 +46,7 @@ public final class GO_Integer extends PropertyType<GO_Integer, Integer> {
      * @param value The value.
      */
     private GO_Integer(final Integer value) {
-        super(value, value.intValue() == 0);
+        super(value, value == 0);
     }
 
     /**
